requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# AI / LLM
google-genai>=1.7.0
langgraph>=0.2.0
# OpenAI SDK — also used for DeepSeek and Groq (OpenAI-compatible APIs)
openai>=1.30.0

# Embeddings (local, no API key needed — default provider)
sentence-transformers>=3.0.0

# Vector store
chromadb>=0.5.0

# Structured store
duckdb>=1.0.0

# Data
pandas>=2.2.0
requests>=2.32.0
pyyaml>=6.0.1

# Utilities
httpx>=0.27.0
